update charter to match updated template
there have been changes to the process section of the charter template (after "Community and Business Group Process") from the discussions in the bluetooth cg and also in some other charter discussions related to the AC.  the older template didn't include GitHub use for issues.  Also, this made it clearer that the formal procedures for decisions and chairs are intended to be rare, only when the more informal way groups work break down.  So, made that clearer and moved rare stuff to the end.