Support opting into supporting reading and writing to all NDEF devices
The most common feedback we have gotten is the lack of support for
the super popular Mifare Classic tags. These are NDEF compatible but
not part of the NFC spec and only supported by NXP reader chips.

This adds an explicit opt-in to reading and writing to these tags

Fixes #160