Compress the Web NFC specific record
The record is now specced better and stored more efficiently
by using short records, no chunking, using a shorter type name
and not storing "https://".

It is also renamed to author record as it is more saying and
more in line with the type. Also before the name Web NFC Message
was used for something that is a record and not a message as
such.