Drop link to potentially malicious example.com web site
It is a very good idea to illustrate the malicious use case with a concrete
example. It does not seem to be such a good idea to link to an external web site
which could exist in theory, and could well be malicious if it existed.

For consistency, this update also drops the link to the legitimate example.com
URL although this link could remain as there are strong guarantees that this
site will always exist for illustrative examples.